merge main branch to this branch

+++ psu_course_review/routers/courses.py
@@ -8,6 +8,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 import math
 
 from .. import models
+from .. import deps
 
 router = APIRouter(prefix="/courses", tags=["courses"])
 
@@ -19,13 +20,18 @@ SIZE_PER_PAGE = 50
 async def create_course(
     course: models.CreatedCourse,
     session: Annotated[AsyncSession, Depends(models.get_session)],
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
 ) -> models.Course:
     db_course = models.DBCourse.model_validate(course)
+
+    db_course.user = current_user
+
     session.add(db_course)
     await session.commit()
     await session.refresh(db_course)
 
     return models.Course.model_validate(db_course)
+
 
 @router.get("")
 async def read_courses(
@@ -54,6 +60,7 @@ async def read_courses(
         )
     )
 
+
 @router.get("/{course_id}")
 async def read_course(
     course_id: int,
@@ -70,11 +77,18 @@ async def update_course(
     course_id: int,
     course: models.UpdatedCourse,
     session: Annotated[AsyncSession, Depends(models.get_session)],
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
 ) -> models.Course:
-    data = course.model_dump()
     db_course = await session.get(models.DBCourse, course_id)
     if db_course is None:
         raise HTTPException(status_code=404, detail="Course not found")
+
+    if db_course.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Forbidden")
+
+    course.user_id = db_course.user_id
+
+    data = course.model_dump()
 
     db_course.sqlmodel_update(data)
 
@@ -84,14 +98,19 @@ async def update_course(
 
     return models.Course.model_validate(db_course)
 
+
 @router.delete("/{course_id}")
 async def delete_course(
     course_id: int,
     session: Annotated[AsyncSession, Depends(models.get_session)],
+    current_user: Annotated[models.User, Depends(deps.get_current_user)],
 ) -> dict:
     db_course = await session.get(models.DBCourse, course_id)
     if db_course is None:
         raise HTTPException(status_code=404, detail="Course not found")
+
+    if db_course.user_id != current_user.id:
+        raise HTTPException(status_code=403, detail="Forbidden")
 
     await session.delete(db_course)
     await session.commit()
